allow enabling/disabling cache
